backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfminer.six
python-docx
requests
python-dotenv
